fix(csv): sanitize import path segments
