+++ src/util.ts
@@ -1,31 +1,5 @@
 import { FrontMatterCache, stringifyYaml } from 'obsidian';
-
-let slashesRe = /[/\\]/g;
-let illegalRe = /[\?<>:\*\|"]/g;
-let controlRe = /[\x00-\x1f\x80-\x9f]/g;
-let reservedRe = /^\.+$/;
-let windowsReservedRe = /^(con|prn|aux|nul|com[0-9]|lpt[0-9])(\..*)?$/i;
-let windowsTrailingRe = /[\. ]+$/;
-let startsWithDotRe = /^\./; // Regular expression to match filenames starting with "."
-let badLinkRe = /[\[\]#|^]/g; // Regular expression to match characters that interferes with links: [ ] # | ^
-
-// First remove illegal characters such as spaces and periods, then check for Windows reserved words.
-export function sanitizeFileName(name: string) {
-	const sanitized = name
-		.replace(slashesRe, '-') // Replace slashes with dash
-		.replace(illegalRe, '')
-		.replace(controlRe, '')
-		.replace(reservedRe, '')
-		.replace(windowsTrailingRe, '')
-		.replace(windowsReservedRe, '')
-		.replace(startsWithDotRe, '')
-		.replace(badLinkRe, '');
-
-	// If the result is empty or only whitespace after sanitization, return a default name
-	// This prevents creating files like ".md" (no name) or folders with only spaces
-	const trimmed = sanitized.trim();
-	return trimmed || 'Untitled';
-}
+export { sanitizeFileName, sanitizeFilePath } from './sanitize';
 
 export function genUid(length: number): string {
 	let array: string[] = [];
